% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{calculateCCFmaxSearch}
\alias{calculateCCFmaxSearch}
\title{Compute maximum correlation for two images, taking into account a sequence of
rotation angles and all integer-valued translations (updated)}
\usage{
calculateCCFmaxSearch(image1, image2, pad = TRUE)
}
\arguments{
\item{image1}{first image matrix}

\item{image2}{second image matrix. The two images do not need to be the same
size.}

\item{pad}{logical value for whether the smaller image should be zero-padded.
The default is TRUE. Only set this to FALSE if the images are of the same
size and are square.}
}
\value{
A list with four items: 1) maximum correlation taking into account
  many possible rotations and translations of the second image, 2) the
  corresponding dx, the horizontal translation, 3) dy, the vertical
  translation, and 4) theta, the corresponding rotation angle.
}
\description{
This is an updated version of calculateCCFmax(). It normalizes the images
after rotation (mean 0, variance 1), and uses a slightly different grid
search to reduce runtime.
}
\examples{
\dontrun{
calculateCCFmaxSearch(processedExample, processedExample2)
}

}
