% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartridges.R
\docType{package}
\name{cartridges}
\alias{cartridges}
\alias{cartridges-package}
\title{cartridges: A package for comparing cartridge case images.}
\description{
This package contains utilities to read and plot cartridge case images from
the NIST Ballistics Toolmark database. It implements an algorithm to compare
cartridge case images after pre-processing. Go to
\url{https://github.com/xhtai/cartridges} for more information.
}
