% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_level.R
\name{centerBFprimer}
\alias{centerBFprimer}
\title{Center image}
\usage{
centerBFprimer(inputImage, primer)
}
\arguments{
\item{inputImage}{1919 x 1919 matrix}

\item{primer}{binary image in \code{EBImage} format, as returned by
\code{findPrimer}}
}
\value{
A list with two items: 1) \code{centeredBF}: A 1769 x 1769 matrix,
  with pixel values ranging from 0-255. Areas that are not part of the
  breechface impression are set to NA. 2) \code{centeredPrimer}: A 1769 x
  1769 binary matrix denoting the primer region, after centering. This is to
  be used in a later step.
}
\description{
Given a 1919 x 1919 image matrix, and its binary primer image (in
\code{EBImage} format), computes the center of the image as the center of the
primer region, and returns a 1769 x 1769 image matrix where the center of the
primer is located at (884, 884).
}
\examples{
\dontrun{
centeredExample <- center(FPexample, primerExample)
}

}

