% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_compare.R
\name{comparison}
\alias{comparison}
\title{Compute correlation for two images}
\usage{
comparison(image1, image2)
}
\arguments{
\item{image1}{first image matrix}

\item{image2}{second image matrix. Note that the two matrices have to be the
same size.}
}
\value{
A list with three items: 1) maximum correlation taking into account
  all possible translations, 2) dx, the horizontal translation corresponding
  to the maximum correlation, and 3) dy, the vertical translation
  corresponding to the maximum correlation.
}
\description{
Given two processed (and possibly rotated) images, compute the
cross-correlation function and return the largest correlation value and its
associated translation parameters.
}
\examples{
\dontrun{
comparison(processedExample, processedExample2)
}

}
