% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6_prob.R
\name{computeProb}
\alias{computeProb}
\title{Compute the probability of obtaining a higher correlation by chance}
\usage{
computeProb(corr, empiricalNull)
}
\arguments{
\item{corr}{maximum correlation between two images, taking into account all
translations and many rotation angles, e.g. the output of
\code{calculateCCFmax()}.}

\item{empiricalNull}{the non-match distribution to compare the value of
\code{corr} against. We suggest using the empirical null distribution based
on all pairwise comparisons of non-matches in a known database.}
}
\value{
the probability of obtaining a higher similarity score by chance.
}
\description{
Given a reference distribution of correlation values of non-matches, compute
the probability of obtaining a higher correlation, under the assumption that
the two images are not a match, i.e. not from the same gun.
}
\examples{
\dontrun{
computeProb(.1, rnorm(50, .07, 1))
}

}

