% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_outlier_filter.R
\name{cropBorders}
\alias{cropBorders}
\title{Crop borders of an image}
\usage{
cropBorders(image, primer)
}
\arguments{
\item{image}{1769 x 1769 matrix}

\item{primer}{A 1769 x 1769 binary matrix denoting the primer region}
}
\value{
A matrix, smaller than 1769 x 1769, containing only the valid primer
  region. Not necessarily square.
}
\description{
Given an image with NA values in the borders, crops out these borders to
leave the smallest rectangular region containing valid data.
}
\examples{
\dontrun{
croppedExample <- cropBorders(removedExample, centeredExample$centeredPrimer)
}

}
