% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_selectBF.R
\name{findFP}
\alias{findFP}
\title{Find and remove firing pin region}
\usage{
findFP(fileName, primer)
}
\arguments{
\item{fileName}{location of image file. This function has been tested on
\code{png} files.}

\item{primer}{binary image in \code{EBImage} format, as returned by
\code{findPrimer}}
}
\value{
A 1919 x 1919 matrix, in the original orientation, i.e. a cropped
  version of the original 1944 x 2592 image.
}
\description{
After finding the primer region using \code{findPrimer()}, we now remove the
firing pin impression, returning a 1919 x 1919 matrix, where areas that are
not part of the breechface impression are set to 0. Pixel values range from 0
to 255. For more information on the steps involved in finding the firing pin
region, refer to the GitHub page.
}
\examples{
\dontrun{
FPexample <- findFP(system.file("extdata", "NBIDE R BF 118.png", package = "cartridges"), primer = primerExample)
}

}

