% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_selectBF.R
\name{findPrimer}
\alias{findPrimer}
\title{Read cartridge case image and find primer region}
\usage{
findPrimer(fileName)
}
\arguments{
\item{fileName}{location of image file. This function has been tested on
\code{png} files.}
}
\value{
A 1919 x 1919 binary image in \code{EBImage} format. Values 1
  indicate the primer region.
}
\description{
Images must be of the standard format as provided in the NIST Ballistics and
Research Toolmark Database. The dimensions are 1944 x 2592 pixels, and images
are on a 255-grayscale. Here we use ring light images. This function finds
the primer region, which is the circular region containing the breechface
marks and firing pin impression. By default images are cropped to 1919 x 1919
pixels. For more information on the steps involved in finding this region,
refer to the GitHub page.
}
\details{
This function uses the \code{EBImage} package, which stores the image as a
2592 x 1944 matrix. The primer region that is returned is similarly an
\code{EBImage} image, where the primer region is stored in a matrix with
binary values. If using this matrix directly on the image matrix in the
original orientation (1944 x 2592), note that the should first be transposed.
}
\examples{
\dontrun{
primerExample <- findPrimer(system.file("extdata", "NBIDE R BF 118.png", package = "cartridges"))
}

}

