% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunctions.R
\name{fitBasis}
\alias{fitBasis}
\title{Fit basis functions}
\usage{
fitBasis(image, basis)
}
\arguments{
\item{image}{centered image. Image has to be square.}

\item{basis}{list of basis functions of the appropriate dimension, such as
those produced by \code{getBasisFunctions()}.}
}
\value{
A data frame with the number of pixels in each basis function, basis
  function coefficient and residual sum of squares for each basis function.
}
\description{
Fit basis functions to a centered image.
}
\examples{
basis3 <- getBasisFunctions(3)
sampleBasis <- fitBasis(matrix(1:9,nrow=3),basis3)
}

