% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_outlier_filter.R
\name{gaussianFilter}
\alias{gaussianFilter}
\title{Filter image.}
\usage{
gaussianFilter(inputImage, nonBF)
}
\arguments{
\item{inputImage}{image matrix}

\item{nonBF}{binary matrix of the same dimensions as \code{inputImage}, where
1s indicate the non-breechface region.}
}
\value{
The image matrix after filtering.
}
\description{
This function applies a low and high-pass Gaussian filter. The low-pass
filter uses an 11 x 11 matrix with \code{sigma = 6}, and the high-pass filter
uses a 149 x 149 matrix with \code{sigma = 44}. This is translated from code
by Joseph Roth and Zach Richardson.
}
\examples{
\dontrun{
# first run this:
croppedExample <- cropBorders(removedExample, centeredExample$centeredPrimer)
# then
nonBF <- is.na(croppedExample)
processedExample <- gaussianFilter(inpaintedExample, nonBF)
}

}

