% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{getClust}
\alias{getClust}
\title{Going from links to clusters}
\usage{
getClust(pairs, preds, hash1, hash2)
}
\arguments{
\item{pairs}{pairs with `preds` column, a binary indicator for whether the
pair is linked after hierarchical clustering}

\item{hash1}{name of column with first item in comparison, e.g. "image1"}

\item{hash2}{name of column with second item in comparison, e.g. "image2"}

\item{pred}{name of column with link, input as character, e.g. "minimax0.5"}
}
\value{
data frame with three columns: `image`, the name of the image,
  `cluster`, the cluster number the image is a member of, and `clusterSize`,
  the size of that cluster
}
\description{
Given pairwise links as generated by `linksAnalysis()`, produce a list of
individual items and their cluster membership
}
