% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_removeCircular.R
\name{getFitted}
\alias{getFitted}
\title{Get fitted image from basis function coefficients}
\usage{
getFitted(basisCoefficients, basis, dimension)
}
\arguments{
\item{basisCoefficients}{vector of basis function coefficients, such as
the \code{coef} column produced by \code{fitBasis()}}

\item{basis}{list of basis functions of the same length as the number of rows
in \code{basisCoefficients}. Can be produced by \code{getBasisFunctions()}.}

\item{dimension}{dimension of output image. This has to match the number of
basis functions in \code{basis}.}
}
\value{
A matrix of pixel values
}
\description{
From basis function coefficients, produce a matrix of pixel values.
}
\examples{
\dontrun{
# first run this:
centeredExample <- center(FPexample, primerExample)
# then
basis <- fitBasis(centeredExample, basis1769)
fitted <- getFitted(basis$fitted, basis1769, 1769)
}
}
