% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{getFittedImage}
\alias{getFittedImage}
\title{Get fitted image from basis function coefficients}
\usage{
getFittedImage(basisCoefficients, basis, dimension)
}
\arguments{
\item{basisCoefficients}{data frame of basis function coefficients, such as
those produced by \code{fitBasis()}}

\item{basis}{list of basis functions of the same length as the number of rows
in \code{basisCoefficients}. Can be produced by \code{getBasisFunctions()}.}

\item{dimension}{dimension of output image. This has to match the number of
basis functions in \code{basis}.}
}
\value{
A matrix of pixel values
}
\description{
From basis function coefficients, produce a matrix of pixel values.
}
\examples{
LL1_3_basis <- fitBasis(LL1_3,basis1919)
fittedImage <- getFittedImage(LL1_3_basis,basis1919,1919)
}

