% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_outlier_filter.R
\name{inpaint_nans}
\alias{inpaint_nans}
\title{Fill in NA values}
\usage{
inpaint_nans(A)
}
\arguments{
\item{A}{image matrix}
}
\value{
A matrix with NA values filled in.
}
\description{
This was translated from John D'Errico's MATLAB code:
\url{https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint-nans},
Copyright (c) 2009, John D'Errico.
}
\examples{
\dontrun{
# first run this:
croppedExample <- cropBorders(removedExample, centeredExample$centeredPrimer)
outlierNAexample <- outlierRejection(croppedExample)
# then
inpaintedExample <- inpaint_nans(outlierNAexample)
}

}
