% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_level.R
\name{levelBF}
\alias{levelBF}
\title{Level image}
\usage{
levelBF(inputImage)
}
\arguments{
\item{inputImage}{1769 x 1769 matrix}
}
\value{
A leveled 1769 x 1769 image matrix.
}
\description{
Given a 1769 x 1769 image matrix, fits a plane through the image and returns
the residuals. Areas that are not part of the breechface impression are set
to NA. Removing residuals means that possible values are -255 to 255.
}
\examples{
\dontrun{
# first run this:
centeredExample <- centerBFprimer(FPexample, primerExample)
# then:
leveledExample <- levelBF(centeredExample$centeredBF)
}

}

