% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{linksAnalysis}
\alias{linksAnalysis}
\title{Hierarchical clustering to generate final clusters}
\usage{
linksAnalysis(pairs, predCol, myCutoff, myMethod, hash1, hash2)
}
\arguments{
\item{pairs}{pairs with `predCol` column, which are similarity
scores/predictions (higher = more similar)}

\item{predCol}{name of column with similarities/predictions, input as
character, e.g. "myPreds"}

\item{myCutoff}{similarity cutoff (>= myCutoff corresponds to a "match")}

\item{myMethod}{type of linkage, can be "single", "complete", "average",
"minimax"}

\item{hash1}{name of column with first item in comparison, e.g. "image1"}

\item{hash2}{name of column with second item in comparison, e.g. "image2"}
}
\value{
vector (logical) indicating if pair is matched or not, length is the
  same as `pairs`
}
\description{
Given similarities/predictions for each pair, generate transitive closures
using hierarchical clustering, returning pairwise match prediction
}
