% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL1_3.R
\docType{data}
\name{LL1_3}
\alias{LL1_3}
\title{Matrix of pixel values of image from Laura Lightstone study.}
\format{A matrix with dimensions 1919 x 1919. Each entry is a pixel value
  from 0 to 255.}
\source{
\url{http://www.nist.gov/forensics/ballisticsdb/index.cfm}
}
\usage{
LL1_3
}
\description{
A matrix containing the pixel values (0-255) from the Laura Lightstone study
in the NIST Ballistics Toolmark Research Database. This image is a 2D
breechface ring light image, and is the third image in the study (third test
fire from the first pistol slide). The original filename in the NIST download
is "ATF SW 1-3 BF R.tif". This image was read in and cropped to 1919 x 1919
pixels using the function \code{\link{readCropTIFF}}.
}
\keyword{datasets}

