% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forClustering.R
\name{longToSquare}
\alias{longToSquare}
\title{Generate distance matrix from pairwise list and distance (or similarity)
column}
\usage{
longToSquare(allPairwise, pairColNums, distSimCol, myDist = TRUE)
}
\arguments{
\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least three columns, one representing the first item
in the comparison, one representing the second item, and the last
representing a similarity or distance metric. These are enumerated in the
next two parameters.}

\item{pairColNums}{vector of length 2 indicating the column numbers in
`allPairwise` of 1. item 1 in comparison, 2. item 2 in comparison}

\item{distSimCol}{name of column in `allPairwise` indicating distances or
similarities, input as character, e.g. "l2dist". If this is a similarity
and not a difference, input `myDist` parameter to be FALSE}

\item{myDist}{is `distSimCol` a distance or similarity measure? Default TRUE,
i.e. distance measure}
}
\value{
symmetric distance matrix
}
\description{
Generate distance matrix from pairwise list and distance (or similarity)
column
}
