% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forClustering.R
\name{makeLinkCol}
\alias{makeLinkCol}
\title{Generate binary column representing links between pairs, given dendrogram and
cut specification}
\usage{
makeLinkCol(allPairwise, pairColNums, htree, numClust, linkage)
}
\arguments{
\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least two columns, one representing the first item in
the comparison, and one representing the second item.}

\item{pairColNums}{vector of length 2 indicating the column numbers of 1.
item 1 in comparison, 2. item 2 in comparison}

\item{htree}{tree (dendrogram) after hierarchical clustering, returned by
`getHcluster`}

\item{numClust}{number of desired clusters in clustering, can be a number
from 1 to n, where n is the number of items to be clustered}

\item{linkage}{one of "single", "complete", "average", "centroid", "minimax"}
}
\value{
binary vector representing whether links exist between pairs in
  `allPairwise`
}
\description{
Generate binary column representing links between pairs, given dendrogram and
cut specification
}
