% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{plotImage}
\alias{plotImage}
\title{Plot an image}
\usage{
plotImage(image, type, grayscale, main)
}
\arguments{
\item{image}{image to be plotted}

\item{type}{either \code{"original"} for images on the
original scale (pixel values 0-255), or \code{"residuals"}
for residual pixel values (-255 to 255).}

\item{grayscale}{logical value indicating whether or not
the grayscale is to be plotted.}

\item{main}{title for plots with grayscale}
}
\description{
Plot an image
}
\examples{
data(LL1_3)
plotImage(LL1_3,"original",grayscale=FALSE)
}

