% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerExample.R
\docType{data}
\name{primerExample}
\alias{primerExample}
\title{Binary image in \code{EBImage} format indicating the primer region}
\format{A 1919 x 1919 binary image in \code{EBImage} format. Values 1
  indicate the primer region.}
\source{
\url{https://tsapps.nist.gov/NRBTD}
}
\usage{
primerExample
}
\description{
Primer region, as returned by \code{findPrimer("NBIDE R BF 118.png")}. This
is part of the first pre-processing step.
}
\keyword{datasets}

