% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{readCartridgeImage}
\alias{readCartridgeImage}
\title{Read in a cartridge image}
\usage{
readCartridgeImage(fileName)
}
\arguments{
\item{fileName}{location of image file to be read in}
}
\value{
A 1944 x 2592 matrix containing pixel values from 0 to 255.
}
\description{
Images from the NIST Ballistics and Research Toolmark Database are 1944 x
2592 pixels, and on a 255-grayscale. This function reads in such images and
returns a matrix of pixel values.
}
\examples{
\dontrun{
exampleImage <- readCartridgeImage(system.file("extdata", "NBIDE R BF 118.png",
 package="cartridges"))
}
}
