% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_removeCircular.R
\name{removeCircular}
\alias{removeCircular}
\title{Remove circular symmetry of an image}
\usage{
removeCircular(inputImage)
}
\arguments{
\item{inputImage}{1769 x 1769 matrix}
}
\value{
A 1769 x 1769 image matrix.
}
\description{
Given a 1769 x 1769 image matrix, fits a circularly symmetric model on the
image and returns the residuals. Areas that are not part of the breechface
impression are set to NA. If this is run after \code{levelBF}, removing
residuals again means that possible values are -510 to 510.
}
\examples{
\dontrun{
# first run this:
centeredExample <- centerBFprimer(FPexample, primerExample)
leveledExample <- levelBF(centeredExample$centeredBF)
# then
removedExample <- removeCircular(leveledExample)
}

}
