% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{removeDups}
\alias{removeDups}
\title{Given a data set with all pairwise comparisons including A-B and B-A
comparisons, take the larger of the similarity score}
\usage{
removeDups(allResults, similarityCol)
}
\arguments{
\item{allResults}{needs to have columns `compare`, `newImage`, and a column
with similarity scores (see next parameter)}

\item{similarityCol}{name of column with similarity scores, e.g. `"corr"`}
}
\value{
a data frame with one row per comparison, adding a column for the B-A
  match and a column `corrMax`
}
\description{
Given a data set with all pairwise comparisons including A-B and B-A
comparisons, take the larger of the similarity score
}
