% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{roughCenter}
\alias{roughCenter}
\title{Rough search for the center of a 1919 x 1919 image}
\usage{
roughCenter(image, basis, startingRow)
}
\arguments{
\item{image}{image that rough center is to be
determined for}

\item{basis}{subset of basis functions of the
appropriate dimension (1919 x 1919). Recommended
restrictions are basis functions with
distances from center of 450-1000
pixels, and with 8 pixels in each basis function.}

\item{startingRow}{row to be searched}
}
\value{
A data frame with 402 rows,
each row with the coordinates of the center
considered, and the maximum number of continuous
zeros.
}
\description{
This method is intended only as a rough search for
a starting pixel location to be used by \code{gridSearch()}.
An appropriate center would fit the ring in the image
well, and the goal is to find a center producing the
longest continuous string of fitted small coefficients where the
ring is expected to be. An approximate location of the
ring is 450-1000 pixels from the center, so only the
subset of basis functions corresponding to this distance
is recommended.
}
\details{
Along an input row (recommended rows are 940 and 980),
we consider 201 possible center locations from columns
860 to 1060. Then after selecting an appropriate column,
we consider 201 possible rows (100 above and 100 below
the input row). We recommend use of a further
subset of basis functions (those with 8 pixels in each
basis function, since these are the most common) for speed.
Also, instead of fitting each basis function (computing
the mean of the pixel values) and searching for the
longest string of small values, we convert the images from
256-grayscale to binary values, with values from 0 to 127
taking the value 0 and those from 128 to 255 taking the
value 1. Now at each possible center, we compute the sum
of pixel values of each basis function, and find the
center resulting in the longest continuous string of zeros.
}
\examples{
data(LL1_3,basis1919)
basis1919_8pixels_rad450_1000 <- subsetBasis(basis1919,c(450,1000),8)
center <- roughCenter(LL1_3, basis1919_8pixels_rad450_1000, 940)
}

