% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{shiftedImage}
\alias{shiftedImage}
\title{Shift a 1919 x 1919 image}
\usage{
shiftedImage(image, centeri, centerj)
}
\arguments{
\item{image}{image that is to be shifted}

\item{centeri}{i-coordinate of desired center}

\item{centerj}{j-coordinate of desired center}
}
\value{
A shifted 1919 x 1919 image. Pixel values
that are not available are set to 0.
}
\description{
Shift a 1919 x 1919 image
}
\examples{
data(LL1_3)
shifted <- shiftedImage(LL1_3, 910, 910)
}

