% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunctions.R
\name{subsetBasis}
\alias{subsetBasis}
\title{Get subset of basis functions}
\usage{
subsetBasis(originalBasis, distFromCenter, numPixels)
}
\arguments{
\item{originalBasis}{list object produced by \code{getBasisFunctions()}}

\item{distFromCenter}{vector of length 2 indicating the minimum and maximum
distances from the center}

\item{numPixels}{integer value indicating the required number of pixels in
each basis function}
}
\value{
A subset of the original basis with basis functions of the required
  distances from the center and number of pixels.
}
\description{
Get a subset of basis functions produced by \code{getBasisFunctions()},
subsetting based on distance from center and the number of pixels in each
basis function.
}
\examples{
basis3 <- getBasisFunctions(3)
basis3_rad1_4pixels <- subsetBasis(basis3,c(1,1),4)

}

