% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{surfacePlot}
\alias{surfacePlot}
\title{Surface plot of total residual sums of squares}
\usage{
surfacePlot(RSS, main)
}
\arguments{
\item{RSS}{data frame of RSS values, in the format produced by
\code{gridSearch()}}

\item{main}{title of plot}
}
\description{
Produces a surface plot of the total residual sum of squares of all the
centers tested
}
\examples{
\dontrun{
basis1919_rad450_1000<-subsetBasis(basis1919,c(450,1000))
RSSdtf<-gridSearch(LL1_3,basis1919_rad450_1000,970,962) # warning: this is
a long computation and may take an hour or more, depending on your system
surfacePlot(RSSdtf,main="Total RSS of Basis Functions \\n
             with Radius of 450-1000 Pixels")
}
}

