% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartridges.R
\docType{package}
\name{cartridges}
\alias{cartridges}
\alias{cartridges-package}
\title{cartridges: A package for computing signatures from cartridge images.}
\description{
This package contains utilities to read, center and plot cartridge images
from the NIST Ballistics Toolmark database. It implements a method of
creating a signature from such images. This is work in progress. Go to
\url{https://github.com/xhtai/cartridges} for more information.
}

