% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{getFittedImage}
\alias{getFittedImage}
\title{Get fitted image from basis function coefficients}
\usage{
getFittedImage(basisCoefficients, basis, dimension)
}
\arguments{
\item{basisCoefficients}{data frame of basis function
coefficients, such as those produced by
\code{fitBasis()}}

\item{basis}{list of basis functions of the appropriate
dimension, such as those
produced by \code{getBasisFunctions()}.}

\item{dimension}{dimension of image}

\item{image}{centered image}
}
\value{
A matrix of pixel values
}
\description{
Get fitted image from basis function coefficients
}
\examples{
LL1_3_basis <- fitBasis(LL1_3,basis1919)
fittedImage <- getFittedImage(LL1_3_basis,basis1919,1919)
}

